fix: Moved kilocode specific code to kilocode folder
