+++ packages/opencode/src/cli/cmd/tui/app.tsx
@@ -28,7 +28,7 @@ import { PromptHistoryProvider } from "./component/prompt/history"
 import { FrecencyProvider } from "./component/prompt/frecency"
 import { PromptStashProvider } from "./component/prompt/stash"
 import { DialogAlert } from "./ui/dialog-alert"
-import { MessageV2 } from "@/session/message-v2" // kilocode_change
+import { isKiloError } from "@/kilocode/kilo-errors" // kilocode_change
 import { ToastProvider, useToast } from "./ui/toast"
 import { ExitProvider, useExit } from "./context/exit"
 import { Session as SessionApi } from "@/session"
@@ -719,7 +719,7 @@ function App() {
     const error = evt.properties.error
     if (error && typeof error === "object" && error.name === "MessageAbortedError") return
     // kilocode_change start - Skip toast for current Kilo errors - Just show inline
-    if (error && typeof error === "object" && MessageV2.isKiloError(error)) return
+    if (error && typeof error === "object" && isKiloError(error)) return
     // kilocode_change end
     const message = (() => {
       if (!error) return "An error occurred"
