+++ packages/opencode/src/session/retry.ts
@@ -1,5 +1,6 @@
 import type { NamedError } from "@opencode-ai/util/error"
 import { MessageV2 } from "./message-v2"
+import { isKiloError } from "@/kilocode/kilo-errors" // kilocode_change
 import { iife } from "@/util/iife"
 
 export namespace SessionRetry {
@@ -63,7 +64,7 @@ export namespace SessionRetry {
     if (MessageV2.ContextOverflowError.isInstance(error)) return undefined
     if (MessageV2.APIError.isInstance(error)) {
       // kilocode_change start - Current Kilo errors require user action (login/signup), don't retry
-      if (MessageV2.isKiloError(error)) return undefined
+      if (isKiloError(error)) return undefined
       // kilocode_change end
       if (!error.data.isRetryable) return undefined
       if (error.data.responseBody?.includes("FreeUsageLimitError"))
